Refactored code to be more universal / accessible
- moved render functions to **render.functions.js**:
 - `selectRenderedPath`
 - `selectStart`
 - `selectPath`
 - `showValidOptions`
 - `createArray` is new function that will generate grid and labyrinth array
