+++ src/store/reducers/defaultSettings.reducer.js
@@ -16,6 +16,7 @@ const initialDefaultSettings = {
     finish: false,
     finishSelected: false,
     finishCanBeSelected: false,
+    generateGrid: false,
     // hours: 0,
     hover: false,
     labyrinthArray: [],
@@ -34,11 +35,13 @@ const initialDefaultSettings = {
     // newCreatedMapIndex: '',
     // ownerSaved: false,
     pathArray: [],
+    pathElement: [],
+    placeWidth: 0,
     // pathSelected: false,
     renderFinish: false,
-    render2DView: false,
     renderVrCreator: false,
     selectedCol: 0,
+    sumOfPaths: 0,
     // seconds: 0,
     startSelected: false,
     time: {},
